+++ packages/orq-rc/src/models/components/conversationwithmessagesresponse.ts
@@ -155,7 +155,7 @@ export const ConversationWithMessagesResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfds3b743v1nrcmv12xfx06h"),
+  _id: z.string().default("conv_01kfdxjhr2n0jfdrd0kcd9nyg0"),
   kind: ConversationWithMessagesResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -189,7 +189,7 @@ export const ConversationWithMessagesResponse$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ConversationWithMessagesResponse
 > = z.object({
-  id: z.string().default("conv_01kfds3b743v1nrcmv12xfx06h"),
+  id: z.string().default("conv_01kfdxjhr2n0jfdrd0kcd9nyg0"),
   kind: ConversationWithMessagesResponseKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),