+++ packages/orq-rc/src/models/operations/createconversationresponse.ts
@@ -337,7 +337,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kfds3b8c7k2drbppappbr3c6"),
+  _id: z.string().default("reasoning_01kfdxjhrs8n7tsx61kr6hy865"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -362,7 +362,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kfds3b8c7k2drbppappbr3c6"),
+  id: z.string().default("reasoning_01kfdxjhrs8n7tsx61kr6hy865"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),
@@ -510,7 +510,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kfds3b7zgmmkay4em87ssjtm"),
+  _id: z.string().default("reasoning_01kfdxjhrj7zexf5q02pdczg6d"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$inboundSchema,
   reasoning: z.string(),
@@ -535,7 +535,7 @@ export const ReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kfds3b7zgmmkay4em87ssjtm"),
+  id: z.string().default("reasoning_01kfdxjhrj7zexf5q02pdczg6d"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$outboundSchema,
   reasoning: z.string(),