+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4703,7 +4703,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KFDS3B4M4G6T640KKK1HCZ86"),
+  id: z.string().default("01KFDXJHNN0Y53K8FJ5X7ZY8JA"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>
@@ -4724,7 +4724,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KFDS3B4M4G6T640KKK1HCZ86"),
+  id: z.string().default("01KFDXJHNN0Y53K8FJ5X7ZY8JA"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>