+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -316,7 +316,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfds3b792yeeb38f945nkaxb"),
+  _id: z.string().default("conv_01kfdxjhr7bqm18b9kr6ry1ehb"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -348,7 +348,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfds3b792yeeb38f945nkaxb"),
+  id: z.string().default("conv_01kfdxjhr7bqm18b9kr6ry1ehb"),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),