+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -202,7 +202,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KFDS3C5T31C42K20AKF16S9S"),
+  _id: z.string().default("01KFDXJJC0DSPGYV13KF2YB52R"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -245,7 +245,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KFDS3C5T31C42K20AKF16S9S"),
+  id: z.string().default("01KFDXJJC0DSPGYV13KF2YB52R"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,