+++ packages/orq-rc/RELEASES.md
@@ -6029,3 +6029,13 @@ Based on:
 - [typescript v4.2.0-rc.49] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.49 - packages/orq-rc
+
+## 2026-01-20 14:42:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.690.0 (2.797.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.50] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.50 - packages/orq-rc